make storage a required injection
