+++ .changeset/cf-auth-injectable-identity.md
@@ -4,4 +4,4 @@
 
 Make the OAuth identity and token storage injectable, and add a shared env-credential resolver
 
-`createOAuthFlow` now takes the consumer's OAuth identity (`clientId`, `consent`, `redirectUri`) on its context, with `storage` optional, so other Cloudflare CLIs can reuse the flow under their own OAuth app. Also adds a shared env→credential resolver (`getAuthFromEnv`, `getAPIToken`, `requireApiToken`).
+`createOAuthFlow` now takes the consumer's OAuth identity (`clientId`, `consent`, `redirectUri`) and token `storage` on its context, so other Cloudflare CLIs can reuse the flow under their own OAuth app and store tokens in their own location/format. Also adds a shared env→credential resolver (`getAuthFromEnv`, `getAPIToken`, `requireApiToken`).
